backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.0.1"
description = ""
authors = ["Tisha"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.7"
django-ninja = "1.3.0"
mysqlclient = "^2.2.7"
celery = "^5.5.0"
django-redis = "^5.4.0"
django-cors-headers = "^4.7.0"
django-filter = "^25.1"
python-dotenv = "^1.1.0"
django-simpleui = "^2025.1.13"
django-storages = "^1.14.6"
python-magic = "^0.4.27"
django-ninja-jwt = "^5.3.5"
django-ninja-extra = "^0.22.6"
pillow = "^11.1.0"
django-ratelimit = "^4.1.0"
httpx = "^0.28.1"
jmespath = "^1.0.1"
psycopg2-binary = "^2.9.10"
django-taggit = "^6.1.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.1.0"
ipython = "^9.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
